+++ src/main.js
@@ -26,7 +26,6 @@ class PlatformIOVSCodeExtension {
     this.pioTerm = undefined;
     this.pioHome = undefined;
 
-    this._isMonitorRun = false;
     this._enterpriseSettings = undefined;
   }
 
@@ -181,13 +180,10 @@ class PlatformIOVSCodeExtension {
     ));
     this.context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.upload',
-      async () => {
-        await this.terminateMonitorTask();
-
+      () => {
         let task = 'PlatformIO: Upload';
         if (this.getConfig().get('forceUploadAndMonitor')) {
           task = 'PlatformIO: Upload and Monitor';
-          this._isMonitorRun = true;
         }
         vscode.commands.executeCommand('workbench.action.tasks.runTask', task);
       }
@@ -198,10 +194,7 @@ class PlatformIOVSCodeExtension {
     ));
     this.context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.test',
-      async () => {
-        await this.terminateMonitorTask();
-        vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Test');
-      }
+      () => vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Test')
     ));
     this.context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.clean',
@@ -209,11 +202,7 @@ class PlatformIOVSCodeExtension {
     ));
     this.context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.serialMonitor',
-      async () => {
-        await this.terminateMonitorTask();
-        this._isMonitorRun = true;
-        vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Monitor');
-      }
+      () => vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Monitor')
     ));
     this.context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.newTerminal',
@@ -227,19 +216,6 @@ class PlatformIOVSCodeExtension {
       'platformio-ide.upgradeCore',
       () => this.pioTerm.sendText('pio upgrade')
     ));
-  }
-
-  async terminateMonitorTask() {
-    if (!this._isMonitorRun) {
-      return;
-    }
-    try {
-      await vscode.commands.executeCommand('workbench.action.tasks.terminate');
-    } catch (err) {
-      console.error(err);
-    }
-    this._isMonitorRun = false;
-    return new Promise(resolve => setTimeout(() => resolve(), 500));
   }
 
   initDebug() {