+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.16.0 (2018-??-??)
 
+**Requires VSCode 1.24 or above**
+
+* Automatically close Serial Port Monitor before uploading/testing (issue [#49](https://github.com/platformio/platformio-vscode-ide/issues/49))
+* Added new configuration option `autoCloseSerialMonitor`, which is set to `true` by default
 * Added "Report a problem" action/button when error occurs
 * Improved PIO Core installer using `pip` as Python module
 
@@ -17,7 +21,7 @@
 
 ## 0.15.0 (2018-05-08)
 
-**Requires VSCode 1.23.0 or above**
+**Requires VSCode 1.23 or above**
 
 * New UI for [PIO Unified Debugger](http://docs.platformio.org/page/plus/debugging.html):
   - Conditional Breakpoints